リファクタリング対応
