--- lib/memo_detail_page.dart
@@ -1,71 +1,0 @@
-import 'package:flutter/cupertino.dart';
-import 'package:flutter/material.dart';
-import 'package:team14/memoTemplate.dart';
-import 'package:team14/spot.dart';
-
-class MemoDetailPage extends StatefulWidget {
-  const MemoDetailPage({Key? key}) : super(key: key);
-
-  @override
-  State<MemoDetailPage> createState() => _MemoDetailPageState();
-}
-
-class _MemoDetailPageState extends State<MemoDetailPage> {
-  MemoTemplate mt = MemoTemplate('template', true, {'駐輪しやすい', '受け取り待機時間なし', '店員の態度がいい'}, {'オファー金額', '800~1000円', '1000~1200円'});
-  Spot spot = Spot(
-      "マクドナルド 石橋店", 21.0, 34.807805, 135.444553, 1, "近くにチェーン店がある", [true, true, true], 1, DateTime.now(), DateTime.now());
-
-  Widget listItem(String leftText, String rightText) {
-    return Container(
-      decoration: const BoxDecoration(
-        border: Border(
-          bottom: BorderSide(color: Colors.black12),
-        ),
-      ),
-      padding: const EdgeInsets.all(10.0),
-      child: Row(
-        children: [
-          Expanded(
-            child: Text(
-              leftText,
-              textAlign: TextAlign.left,
-            ),
-          ),
-          Expanded(
-            child: Text(
-              rightText,
-              textAlign: TextAlign.right,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        flexibleSpace: const Text(
-          'メモ詳細',
-          style: TextStyle(fontSize: 30.0, color: Colors.white),
-        ),
-      ),
-      body: Container(
-        child: ListView(
-          children: [
-            listItem('タイトル', spot.title),
-            listItem('位置情報', '${spot.gpsLatitude}, ${spot.gpsLongitude}'),
-            listItem('気温', '${spot.temperature}'),
-            if (mt.textBox)
-              listItem('テキスト', '${spot.textBox}'),
-            for (var i=0; i<mt.multipleSelectList.length; i++)
-              listItem(mt.multipleSelectList.elementAt(i), spot.multipleSelectList![i] ? 'はい' : 'いいえ'),
-            if (mt.singleSelectList.isNotEmpty)
-              listItem(mt.singleSelectList.first, mt.singleSelectList.elementAt(spot.singleSelect!+1)),
-          ],
-        ),
-      ),
-    );
-  }
-}
